+++ src/kotori.node/kotori/node/udp.py
@@ -20,7 +20,7 @@ class UdpPublisher(ApplicationSession):
     - https://github.com/tavendo/AutobahnPython/blob/master/examples/twisted/wamp/session/fromoutside/client.py
     """
 
-    @inlineCallbacks
+    #@inlineCallbacks
     def onJoin(self, details):
         print("session attached")
 
@@ -30,14 +30,14 @@ class UdpPublisher(ApplicationSession):
 
     def onDisconnect(self):
         print("disconnected")
-        reactor.stop()
+        #reactor.stop()
 
 
 def run_wamp_client():
     #log.startLogging(sys.stdout)
 
     # connect to crossbar router/broker
-    runner = ApplicationRunner("ws://localhost:9000/ws", "kotori-realm")
+    runner = ApplicationRunner(u'ws://localhost:9000/ws', u'kotori-realm')
     #runner = ApplicationRunner("ws://master.example.com:9000/ws", "kotori-realm")
 
     runner.run(UdpPublisher, start_reactor=False)